pyproject.toml
[tool.poetry]
name = "goshtdokoni-bot"
version = "0.1.0"
description = ""
authors = ["Tekindev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
aiogram = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
